# Conflicts:
#	src/main/java/com/otd/otd_challenge/application/challenge/ChallengeService.java
#	src/main/java/com/otd/otd_user/entity/User.java

+++ src/main/java/com/otd/otd_challenge/entity/ChallengeDefinition.java
@@ -19,7 +19,7 @@ public class ChallengeDefinition {
     private String cdName;
 
     @Column(nullable = false)
-    private Long cdReward;
+    private int cdReward;
 
     @Column(nullable = false, length = 12)
     private String cdType;
